% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1d.R
\docType{class}
\name{data1d-class}
\alias{data1d}
\alias{data1d-class}
\title{data1d class}
\description{
Class that provides standard methods for working with 1d data item from CCHIC.
Receives a reference to the full data set and then extracts and manipulates the field of interest.
}

