% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2d.R
\docType{class}
\name{data2d-class}
\alias{data2d}
\alias{data2d-class}
\title{data2d class}
\description{
Class that provides standard methods for working with 2d data item from CCHIC
Receives a reference to the full data set and then extracts and manipulates the field of interest.
}

